Split response conversion
